% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data.R
\name{create_wikidk}
\alias{create_wikidk}
\title{Creates Wikipedia dataset}
\usage{
create_wikidk(file)
}
\arguments{
\item{Path}{to file.}
}
\value{
RData file to subdirectory /data.
}
\description{
Run this to recreate wikidk.RData.
}
\examples{
create_wikidk(file = "data/wikipedia-dk-abstract.txt")
}
