% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikidk-data.R
\docType{data}
\name{wikidk}
\alias{wikidk}
\title{Danish Wikipedia dataset}
\format{An object of class \code{character} of length 186125.}
\usage{
data(wikidk)
}
\description{
Run this to recreate wikidk.RData.
}
\examples{
data(wikidk)
}
\keyword{datasets}
